Unified glTF2-based mesh data format across loaders - first pass (#12)
Intention is that an app should not need to care which format it is loading, all loaders in the "same category" should return "normalized" data that the app can work on (e.g. instantiate using preferred 3D library) without additional conditional logic.

* Adds a new doc page that introduces the concept of "categories" of loaders and describes a unified glTF2-inspired format for "mesh and point cloud" category loaders.

* Adds utils to help loaders convert their attributes to the expected formats.

* Updates PLY, PCD, OBJ loaders.

TODO - Second pass to decide if additional standardized format needs to be provided e.g. such as `accessors`.