las: parse two-bytes colors
Before this commit, colors were assumed to be encoded on one byte.
However, the (LAS specification says otherwise, at least since 1.3)[0]
(description of Point Data Record Format 2):
>  Red, Green, Blue values should always be normalized to 16 bit values.

As final colors are stored in a Uint8Array, a division by 256 has been
added. However, to ensure retrocompability, this division is done only
if a value of R, G or B is greater than 255. It is assumed that their
won't be a case of colors being scaled in the 0-255 range, even for
10 bits or greater range of camera color.

[0]: http://www.asprs.org/a/society/committees/standards/asprs_las_spec_v13.pdf#page=11